// src/model/RatingModel.ts
export interface RatingResponse {
  id: string;
  subProductId: string;
  productId: string;
  orderId: string;
  orderProductId: string;
  rating: number;
  comment: string;
  imageUrls: string[];
  userId: string;
  createdAt: string;
  updatedAt: string;
  created: string;
  updated: string;
  avatar: string;
  name: string;
  reply: string
}